refactor: restructure part actions into separate modules
