+++ partcad/src/partcad/actions/part/add.py
@@ -1,0 +1,17 @@
+from pathlib import Path
+from typing import Optional
+import partcad.logging as pc_logging
+from partcad.project import Project
+
+def add_part_action(project: Project, kind: str, path: str, config: Optional[dict] = None):
+    """Add an existing part to the project without copying."""
+    config = config or {}
+    path = Path(path)
+    name = path.stem
+
+    pc_logging.info(f"Adding part '{name}' ({kind}) from '{path}'")
+
+    with pc_logging.Process("AddPart", project.name):
+        result = project.add_part(kind, str(path), config)
+        if result:
+            pc_logging.info(f"Part '{name}' added successfully.")
